moved all state updates for the landing page to home so that they won't be fetched again.
